requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0

sqlalchemy==2.0.29
pymysql==1.1.0

python-jose==3.3.0
passlib[bcrypt]==1.7.4

pydantic==2.6.4
python-dotenv==1.0.1

email-validator==2.1.1
